// apps/admin-panel/src/app/models/IOffer.ts
export interface IOffer{
  id: number;
  companyName: string;
  category: string;
  logo: string;
  logoName: string;
  documentName: string;
  logoUrl: string;
  photo: string;
  photoUrl: string;
  text: string;
}

// {
//     "id": 0,
//     "companyName": "string",
//     "category": "string",
//     "logo": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
//     "photo": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
//     "lastChangeDate": "2020-06-26T13:26:03.059Z"
// }